ilog.c: 6: op: MISRA.ASSIGN.ETYPE.2012
Rule 10.3, assignment to object of narrower essential type or different
essential type category - 2 violations fixed, IDs 987 and 988, tfa
comment: fixing existing code may be counter-productive and introduce bugs

+++ plat/ti/k3low/common/drivers/pm/misc/ilog.c
@@ -38,7 +38,7 @@ int ilog32(uint32_t _v)
 	int ret;
 	int m;
 
-	ret = _v > 0;
+	ret = (int)(_v > 0);
 	m = (_v > 0xFFFFU)<<4;
 	_v >>  = m;
 	ret |= m;
@@ -57,7 +57,7 @@ int ilog32(uint32_t _v)
 #else
 	int ret;
 
-	ret = _v > 0;
+	ret = (int)(_v > 0);
 	_v |= _v >> 1;
 	_v |= _v >> 2;
 	_v |= _v >> 4;
